% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_year_data}
\alias{get_year_data}
\title{Get all baby names data for a specific year}
\usage{
get_year_data(year)
}
\arguments{
\item{year}{The year to filter by. First year available is 2008.}
}
\description{
The top 100 names for a specific year
}
\examples{
\dontrun{
get_year_data(2008)
}

}
